// clients/admin-ui/src/types/api/models/AccessPackageCategory.ts
// This file is auto-generated by @hey-api/openapi-ts

import type { AccessPackageEntry } from "./AccessPackageEntry";

/**
 * AccessPackageCategory
 *
 * A data category grouping within a data use section.
 */
export type AccessPackageCategory = {
  /**
   * Fides Key
   */
  fides_key: string;
  /**
   * Name
   */
  name: string;
  /**
   * Entries
   */
  entries?: Array<AccessPackageEntry>;
};